Query analysis pr comments v4 (#77)
* changed error message
* Refactor function name
* Refactor log statements
* Removed unnecessary comments
* changed the sql server function flow and unit test cases
* resolve lint issues
* Add documentation link for queries
* Removed connection and instance folders
* Change Detect metric function
* Refactor variable name
