+++ src/queryanalysis/utils/utils.go
@@ -8,6 +8,10 @@ import (
 	"strconv"
 	"strings"
 
+	"github.com/newrelic/nri-mssql/src/connection"
+	"github.com/newrelic/nri-mssql/src/instance"
+	"github.com/newrelic/nri-mssql/src/metrics"
+
 	"github.com/jmoiron/sqlx"
 	"github.com/newrelic/infra-integrations-sdk/v3/data/metric"
 	"github.com/newrelic/infra-integrations-sdk/v3/integration"
@@ -15,8 +19,6 @@ import (
 
 	"github.com/newrelic/nri-mssql/src/args"
 	"github.com/newrelic/nri-mssql/src/queryanalysis/config"
-	"github.com/newrelic/nri-mssql/src/queryanalysis/connection"
-	"github.com/newrelic/nri-mssql/src/queryanalysis/instance"
 	"github.com/newrelic/nri-mssql/src/queryanalysis/models"
 )
 
@@ -241,7 +243,7 @@ func IngestQueryMetrics(results []interface{}, queryDetailsDto models.QueryDetai
 		// Iterate over the map and add each key-value pair as a metric
 		for key, value := range resultMap {
 			strValue := fmt.Sprintf("%v", value) // Convert the value to a string representation
-			metricType := DetectMetricType(strValue)
+			metricType := metrics.DetectMetricType(strValue)
 			if metricType == metric.GAUGE {
 				handleGaugeMetric(key, strValue, metricSet)
 			} else {
@@ -257,13 +259,6 @@ func IngestQueryMetrics(results []interface{}, queryDetailsDto models.QueryDetai
 		return err
 	}
 	return nil
-}
-
-func DetectMetricType(value string) metric.SourceType {
-	if _, err := strconv.ParseFloat(value, 64); err != nil {
-		return metric.ATTRIBUTE
-	}
-	return metric.GAUGE
 }
 
 func AnonymizeQueryText(query *string) {
