+++ src/queryanalysis/validation/helper_test.go
@@ -5,8 +5,9 @@ import (
 	"regexp"
 	"testing"
 
+	"github.com/newrelic/nri-mssql/src/connection"
+
 	"github.com/jmoiron/sqlx"
-	"github.com/newrelic/nri-mssql/src/queryanalysis/connection"
 	"github.com/stretchr/testify/assert"
 	"gopkg.in/DATA-DOG/go-sqlmock.v1"
 )
