+++ src/queryanalysis/validation/sql_server_version.go
@@ -1,24 +1,24 @@
 package validation
 
 import (
-	"errors"
 	"fmt"
 	"regexp"
 
+	"github.com/newrelic/nri-mssql/src/connection"
+
 	"github.com/blang/semver/v4"
 	"github.com/newrelic/infra-integrations-sdk/v3/log"
-	"github.com/newrelic/nri-mssql/src/queryanalysis/connection"
 )
 
 const (
 	versionRegexPattern      = `\b(\d+\.\d+\.\d+)\b`
 	getSQLServerVersionQuery = "SELECT @@VERSION"
+	lastSupportedVersion     = 16
+	firstSupportedVersion    = 14
 )
 
 var (
-	versionRegex          = regexp.MustCompile(versionRegexPattern)
-	errEmptyServerVersion = errors.New("server version is empty")
-	errParseVersion       = errors.New("could not parse version from server version string")
+	versionRegex = regexp.MustCompile(versionRegexPattern)
 )
 
 func getSQLServerVersion(sqlConnection *connection.SQLConnection) (string, error) {
@@ -32,18 +32,12 @@ func getSQLServerVersion(sqlConnection *connection.SQLConnection) (string, error
 	if err := rows.Scan(&serverVersion); err != nil {
 		return "", fmt.Errorf("error scanning server version: %w", err)
 	}
-	if serverVersion == "" {
-		return "", fmt.Errorf("%w", errEmptyServerVersion)
-	}
 	log.Debug("Server version: %s", serverVersion)
 	return serverVersion, nil
 }
 
 func parseSQLServerVersion(serverVersion string) (semver.Version, error) {
 	versionStr := versionRegex.FindString(serverVersion)
-	if versionStr == "" {
-		return semver.Version{}, fmt.Errorf("%w", errParseVersion)
-	}
 	log.Debug("Parsed version string: %s", versionStr)
 	version, err := semver.ParseTolerant(versionStr)
 	if err != nil {
@@ -51,17 +45,6 @@ func parseSQLServerVersion(serverVersion string) (semver.Version, error) {
 	}
 	log.Debug("Parsed semantic version: %s", version)
 	return version, nil
-}
-
-func isSQLServerVersionSupported(version semver.Version) bool {
-	supportedVersions := []uint64{16, 15, 14} // Corresponding to SQL Server 2022, 2019, and 2017
-	for _, supportedVersion := range supportedVersions {
-		if version.Major == supportedVersion {
-			return true
-		}
-	}
-	log.Error("Unsupported SQL Server version: %s", version.String())
-	return false
 }
 
 func checkSQLServerVersion(sqlConnection *connection.SQLConnection) (bool, error) {
@@ -73,5 +56,5 @@ func checkSQLServerVersion(sqlConnection *connection.SQLConnection) (bool, error
 	if err != nil {
 		return false, err
 	}
-	return isSQLServerVersionSupported(version), nil
+	return version.Major >= firstSupportedVersion && version.Major <= lastSupportedVersion, nil
 }
